Fix warning: CA1901 P/Invoke declarations should be portable    As it is declared in your code, the return type of P/Invoke 'ExListView.NativeMethods.SendMessage(HandleRef, int, IntPtr, ref ExListView.NativeMethods.LVGROUP)' will be 4 bytes wide on 64-bit platforms. This is not correct, as the actual native declaration of this API indicates it should be 8 bytes wide on 64-bit platforms. Consult the MSDN Platform SDK documentation for help determining what data type should be used instead of 'int'.  GitUI   ExListView.cs   70

+++ GitUI/UserControls/ExListView.cs
@@ -67,7 +67,7 @@ namespace GitUI.UserControls
                                                    ref LVHITTESTINFO lParam);
 
             [DllImport("user32", CharSet = CharSet.Auto)]
-            public static extern int SendMessage(HandleRef hWnd,
+            public static extern IntPtr SendMessage(HandleRef hWnd,
                                                    int msg,
                                                    IntPtr wParam,
                                                    ref LVGROUP lParam);
